+++ packages/driver/cypress/e2e/commands/waiting.cy.js
@@ -61,15 +61,14 @@ describe('src/cy/commands/waiting', () => {
         const response = { foo: 'foo' }
 
         cy
-        .server()
-        .route('GET', /.*/, response).as('fetch')
+        .intercept('GET', /.*/, response).as('fetch')
         .window().then((win) => {
           xhrGet(win, '/foo')
 
           return null
         })
         .wait('@fetch.response').then((xhr) => {
-          expect(xhr.responseBody).to.deep.eq(response)
+          expect(xhr.response.body).to.deep.eq(response)
         })
       })
 
@@ -83,11 +82,13 @@ describe('src/cy/commands/waiting', () => {
         }))
 
         cy
-        .server({ delay: 1000 })
-        .route(/users/, {}).as('getUsers')
+        .intercept(/users/, {
+          body: {},
+          delay: 1000,
+        }).as('getUsers')
         .wait('@getUsers.request').then((xhr) => {
-          expect(xhr.url).to.include('/users')
-          expect(xhr.response).to.be.null
+          expect(xhr.request.url).to.include('/users')
+          expect(xhr.response).to.be.undefined
         })
       })
 
@@ -99,8 +100,7 @@ describe('src/cy/commands/waiting', () => {
         })
 
         cy
-        .server()
-        .route('GET', /.*/, {}).as('fetch')
+        .intercept('GET', /.*/, {}).as('fetch')
         .wait('@fetch', { timeout: 900 })
       })
 
@@ -114,8 +114,7 @@ describe('src/cy/commands/waiting', () => {
         cy.on('command:retry', retry)
 
         cy
-        .server()
-        .route('GET', /.*/, {}).as('fetch')
+        .intercept('GET', /.*/, {}).as('fetch')
         .wait('@fetch').then(() => {
           expect(cy.timeout()).to.eq(prevTimeout)
         })
@@ -131,8 +130,7 @@ describe('src/cy/commands/waiting', () => {
         })
 
         cy
-        .server()
-        .route('GET', '*', {}).as('fetch')
+        .intercept('GET', '*', {}).as('fetch')
         .wait('@fetch').then(() => {
           expect(cy.timeout()).to.eq(199)
         })
@@ -147,8 +145,7 @@ describe('src/cy/commands/waiting', () => {
         })
 
         cy
-        .server()
-        .route('GET', '*', {}).as('fetch')
+        .intercept('GET', '*', {}).as('fetch')
         .wait('@fetch', { requestTimeout: 199 })
       })
 
@@ -156,14 +153,18 @@ describe('src/cy/commands/waiting', () => {
         responseTimeout: 299,
       }, (done) => {
         cy.on('command:retry', (options) => {
-          expect(options.timeout).to.eq(299)
+          if (options.type === 'response') {
+            expect(options.timeout).to.eq(299)
 
-          done()
+            done()
+          }
         })
 
         cy
-        .server({ delay: 100 })
-        .route('GET', '*', {}).as('fetch')
+        .intercept('GET', '*', {
+          body: {},
+          delay: 100,
+        }).as('fetch')
         .window().then((win) => {
           xhrGet(win, '/foo')
 
@@ -174,15 +175,19 @@ describe('src/cy/commands/waiting', () => {
 
       it('waits for responseTimeout override', (done) => {
         cy.on('command:retry', (options) => {
-          expect(options.type).to.eq('response')
-          expect(options.timeout).to.eq(299)
+          if (options.type === 'response') {
+            expect(options.type).to.eq('response')
+            expect(options.timeout).to.eq(299)
 
-          done()
+            done()
+          }
         })
 
         cy
-        .server({ delay: 100 })
-        .route('GET', '*', {}).as('fetch')
+        .intercept('GET', '*', {
+          body: {},
+          delay: 100,
+        }).as('fetch')
         .window().then((win) => {
           xhrGet(win, '/foo')
 
@@ -198,7 +203,7 @@ describe('src/cy/commands/waiting', () => {
           retryCount++
           if (retryCount === 1) {
             expect(options.type).to.eq('request')
-            expect(options.timeout).to.eq(100)
+            expect(options.timeout).to.eq(1000)
 
             // trigger request to move onto response timeout verification
             const win = cy.state('window')
@@ -206,26 +211,27 @@ describe('src/cy/commands/waiting', () => {
             xhrGet(win, '/foo')
           }
 
-          if (retryCount === 2) {
+          if (options.type === 'response') {
             expect(options.type).to.eq('response')
-            expect(options.timeout).to.eq(299)
+            expect(options.timeout).to.eq(1001)
 
             done()
           }
         })
 
         cy
-        .server({ delay: 100 })
-        .route('GET', '*', {}).as('fetch')
-        .wait('@fetch', { requestTimeout: 100, responseTimeout: 299 })
+        .intercept('GET', '*', {
+          body: {},
+          delay: 100,
+        }).as('fetch')
+        .wait('@fetch', { requestTimeout: 1000, responseTimeout: 1001 })
       })
 
       // https://github.com/cypress-io/cypress/issues/369
       it('does not mutate 2nd route methods when using shorthand route', () => {
         cy
-        .server()
-        .route('POST', /foo/, {}).as('getFoo')
-        .route(/bar/, {}).as('getBar')
+        .intercept('POST', /foo/, {}).as('getFoo')
+        .intercept(/bar/, {}).as('getBar')
         .window().then((win) => {
           win.$.post('/foo')
           xhrGet(win, '/bar')
@@ -269,8 +275,7 @@ describe('src/cy/commands/waiting', () => {
             done()
           })
 
-          cy.server()
-          cy.route('GET', /.*/, {}).as('fetch')
+          cy.intercept('GET', /.*/, {}).as('fetch')
           cy.wait('@fetch')
         })
 
@@ -284,8 +289,7 @@ describe('src/cy/commands/waiting', () => {
           })
 
           cy
-          .server()
-          .route(/foo/, {}).as('foo')
+          .intercept(/foo/, {}).as('foo')
           .wait('@foo.request')
         })
 
@@ -297,8 +301,7 @@ describe('src/cy/commands/waiting', () => {
           })
 
           cy
-          .server()
-          .route('*', {}).as('getAny')
+          .intercept('*', {}).as('getAny')
           .wait('getAny').then(() => {})
         })
 
@@ -310,8 +313,7 @@ describe('src/cy/commands/waiting', () => {
           })
 
           cy
-          .server()
-          .route(/foo/, {}).as('foo')
+          .intercept(/foo/, {}).as('foo')
           .window().then((win) => {
             xhrGet(win, '/foo')
 
@@ -328,9 +330,8 @@ describe('src/cy/commands/waiting', () => {
           })
 
           cy
-          .server()
-          .route(/foo/, {}).as('foo')
-          .route(/bar/, {}).as('bar')
+          .intercept(/foo/, {}).as('foo')
+          .intercept(/bar/, {}).as('bar')
           .window().then((win) => {
             xhrGet(win, '/foo')
 
@@ -348,8 +349,7 @@ describe('src/cy/commands/waiting', () => {
           })
 
           cy
-          .server()
-          .route(/foo/, {}).as('foo')
+          .intercept(/foo/, {}).as('foo')
           .get('body').as('bar')
           .window().then((win) => {
             xhrGet(win, '/foo')
@@ -376,9 +376,8 @@ describe('src/cy/commands/waiting', () => {
           })
 
           cy
-          .server()
-          .route(/foo/, {}).as('foo')
-          .route(/bar/, {}).as('bar')
+          .intercept(/foo/, {}).as('foo')
+          .intercept(/bar/, {}).as('bar')
           .wait(['@foo', '@bar'])
         })
 
@@ -400,9 +399,8 @@ describe('src/cy/commands/waiting', () => {
           }))
 
           cy
-          .server()
-          .route(/foo/, { foo: 'foo' }).as('foo')
-          .route(/bar/, { bar: 'bar' }).as('bar')
+          .intercept(/foo/, { foo: 'foo' }).as('foo')
+          .intercept(/bar/, { bar: 'bar' }).as('bar')
           .wait(['@foo', '@bar'])
         })
 
@@ -424,9 +422,8 @@ describe('src/cy/commands/waiting', () => {
           }))
 
           cy
-          .server()
-          .route(/foo/, { foo: 'foo' }).as('foo')
-          .route(/bar/, { bar: 'bar' }).as('bar')
+          .intercept(/foo/, { foo: 'foo' }).as('foo')
+          .intercept(/bar/, { bar: 'bar' }).as('bar')
           .wait(['@foo', '@bar'])
         })
 
@@ -442,9 +439,8 @@ describe('src/cy/commands/waiting', () => {
           })
 
           cy
-          .server()
-          .route(/foo/, {}).as('foo')
-          .route(/bar/, {}).as('bar')
+          .intercept(/foo/, {}).as('foo')
+          .intercept(/bar/, {}).as('bar')
           .wait(['@foo', 'bar'])
         })
 
@@ -461,8 +457,7 @@ describe('src/cy/commands/waiting', () => {
           })
 
           cy
-          .server()
-          .route(/foo/, {}).as('foo')
+          .intercept(/foo/, {}).as('foo')
           .get('body').as('bar')
           .wait(['@foo', '@bar'])
         })
@@ -484,9 +479,8 @@ describe('src/cy/commands/waiting', () => {
           })
 
           cy
-          .server()
-          .route(/foo/, { foo: 'foo' }).as('foo')
-          .route(/bar/, { bar: 'bar' }).as('bar')
+          .intercept(/foo/, { foo: 'foo' }).as('foo')
+          .intercept(/bar/, { bar: 'bar' }).as('bar')
           .wait(['@foo', '@bar'])
         })
 
@@ -515,8 +509,7 @@ describe('src/cy/commands/waiting', () => {
             return xhrGet(win, `/users?num=${response}`)
           })
 
-          cy.server()
-          cy.route(/users/, resp).as('get.users')
+          cy.intercept(/users/, resp).as('get.users')
           cy.wait(['@get.users', '@get.users', '@get.users'])
         })
 
@@ -541,8 +534,7 @@ describe('src/cy/commands/waiting', () => {
           }))
 
           cy
-          .server()
-          .route(/users/, resp).as('getUsers')
+          .intercept(/users/, resp).as('getUsers')
           .wait('@getUsers')
           .wait('@getUsers')
         })
@@ -568,8 +560,7 @@ describe('src/cy/commands/waiting', () => {
           }))
 
           cy
-          .server()
-          .route(/users/, resp).as('getUsers')
+          .intercept(/users/, resp).as('getUsers')
           .wait('@getUsers.request')
           .wait('@getUsers.request')
         })
@@ -584,8 +575,7 @@ describe('src/cy/commands/waiting', () => {
           })
 
           cy
-          .server()
-          .route('*').as('response')
+          .intercept('*').as('response')
           .window().then((win) => {
             xhrGet(win, '/timeout?ms=500')
 
@@ -604,8 +594,7 @@ describe('src/cy/commands/waiting', () => {
           })
 
           cy
-          .server()
-          .route('*').as('response')
+          .intercept('*').as('response')
           .window().then((win) => {
             xhrGet(win, '/timeout?ms=0')
             xhrGet(win, '/timeout?ms=5000')
@@ -625,9 +614,8 @@ describe('src/cy/commands/waiting', () => {
           })
 
           cy
-          .server()
-          .route('/timeout?ms=0').as('foo')
-          .route('/timeout?ms=5000').as('bar')
+          .intercept('/timeout?ms=0').as('foo')
+          .intercept('/timeout?ms=5000').as('bar')
           .window().then((win) => {
             xhrGet(win, '/timeout?ms=0')
             xhrGet(win, '/timeout?ms=5000')
@@ -655,11 +643,13 @@ describe('src/cy/commands/waiting', () => {
           })
 
           cy
-          .server({ delay: 200 })
-          .route(/users/, {}).as('getUsers')
+          .intercept(/users/, {
+            body: {},
+            delay: 200,
+          }).as('getUsers')
           .wait('@getUsers.request').then((xhr) => {
-            expect(xhr.url).to.include('/users')
-            expect(xhr.response).to.be.null
+            expect(xhr.request.url).to.include('/users')
+            expect(xhr.response).to.be.undefined
           })
           .wait('@getUsers')
         })
@@ -686,10 +676,9 @@ describe('src/cy/commands/waiting', () => {
             done()
           })
 
-          cy.server()
-          cy.route('/timeout?ms=2001').as('getOne')
-          cy.route('/timeout?ms=2002').as('getTwo')
-          cy.route(/three/, {}).as('get.three')
+          cy.intercept('/timeout?ms=2001').as('getOne')
+          cy.intercept('/timeout?ms=2002').as('getTwo')
+          cy.intercept(/three/, {}).as('get.three')
           cy.wait(['@getOne', '@getTwo', '@get.three'])
         })
 
@@ -700,7 +689,7 @@ describe('src/cy/commands/waiting', () => {
           const win = cy.state('window')
 
           cy.on('command:retry', (options) => {
-            if (/getThree/.test(options.error)) {
+            if (/getThree/.test(options.error) && options.type === 'response') {
               options._runnableTimeout = 0
             }
           })
@@ -712,10 +701,9 @@ describe('src/cy/commands/waiting', () => {
           })
 
           cy
-          .server()
-          .route('/timeout?ms=1').as('getOne')
-          .route('/timeout?ms=2').as('getTwo')
-          .route('/timeout?ms=3000').as('getThree')
+          .intercept('/timeout?ms=1').as('getOne')
+          .intercept('/timeout?ms=2').as('getTwo')
+          .intercept('/timeout?ms=3000').as('getThree')
           .then(() => {
             xhrGet(win, '/timeout?ms=1')
             xhrGet(win, '/timeout?ms=2')
@@ -758,9 +746,8 @@ describe('src/cy/commands/waiting', () => {
         const resp1 = { foo: 'foo' }
         const resp2 = { bar: 'bar' }
 
-        cy.server()
-        cy.route(/users/, resp1).as('getUsers')
-        cy.route(/posts/, resp2).as('get.posts')
+        cy.intercept(/users/, resp1).as('getUsers')
+        cy.intercept(/posts/, resp2).as('get.posts')
         cy.window().then((win) => {
           xhrGet(win, '/users')
           xhrGet(win, '/posts')
@@ -769,8 +756,8 @@ describe('src/cy/commands/waiting', () => {
         })
 
         cy.wait(['@getUsers', '@get.posts']).spread((xhr1, xhr2) => {
-          expect(xhr1.responseBody).to.deep.eq(resp1)
-          expect(xhr2.responseBody).to.deep.eq(resp2)
+          expect(xhr1.response.body).to.deep.eq(resp1)
+          expect(xhr2.response.body).to.deep.eq(resp2)
         })
       })
     })
@@ -795,19 +782,18 @@ describe('src/cy/commands/waiting', () => {
         })
 
         cy
-        .server()
-        .route(/users/, resp).as('getUsers')
+        .intercept(/users/, resp).as('getUsers')
         .wait('@getUsers').then((xhr) => {
-          expect(xhr.url).to.include('/users?num=1')
-          expect(xhr.responseBody).to.deep.eq(resp)
+          expect(xhr.request.url).to.include('/users?num=1')
+          expect(xhr.response.body).to.deep.eq(resp)
         })
         .wait('@getUsers').then((xhr) => {
-          expect(xhr.url).to.include('/users?num=2')
-          expect(xhr.responseBody).to.deep.eq(resp)
+          expect(xhr.request.url).to.include('/users?num=2')
+          expect(xhr.response.body).to.deep.eq(resp)
         })
         .wait('@getUsers').then((xhr) => {
-          expect(xhr.url).to.include('/users?num=3')
-          expect(xhr.responseBody).to.deep.eq(resp)
+          expect(xhr.request.url).to.include('/users?num=3')
+          expect(xhr.response.body).to.deep.eq(resp)
         })
       })
 
@@ -829,17 +815,16 @@ describe('src/cy/commands/waiting', () => {
         })
 
         cy
-        .server()
-        .route(/users/, resp).as('getUsers')
+        .intercept(/users/, resp).as('getUsers')
         .wait(['@getUsers', '@getUsers', '@getUsers']).spread((xhr1, xhr2, xhr3) => {
-          expect(xhr1.url).to.include('/users?num=1')
+          expect(xhr1.request.url).to.include('/users?num=1')
           expect(xhr2.url).to.include('/users?num=2')
 
-          expect(xhr3.url).to.include('/users?num=3')
+          expect(xhr3.request.url).to.include('/users?num=3')
         }).wait('@getUsers').then((xhr) => {
-          expect(xhr.url).to.include('/users?num=4')
+          expect(xhr.request.url).to.include('/users?num=4')
 
-          expect(xhr.responseBody).to.deep.eq(resp)
+          expect(xhr.response.body).to.deep.eq(resp)
         })
       })
 
@@ -881,6 +866,10 @@ describe('src/cy/commands/waiting', () => {
 
         cy.on('log:added', (attrs, log) => {
           this.lastLog = log
+          if (log.get('name') === 'wait') {
+            this.lastWaitLog = log
+          }
+
           this.logs.push(log)
         })
 
@@ -994,8 +983,7 @@ describe('src/cy/commands/waiting', () => {
             done()
           })
 
-          cy.server()
-          cy.route(/foo/, {}).as('getFoo')
+          cy.intercept(/foo/, {}).as('getFoo')
           cy.noop({}).wait('@getFoo')
         })
 
@@ -1014,9 +1002,9 @@ describe('src/cy/commands/waiting', () => {
           requestTimeout: 100,
         }, function (done) {
           cy.on('fail', (err) => {
-            const { lastLog } = this
+            const { lastWaitLog } = this
 
-            expect(lastLog.get('error')).to.eq(err)
+            expect(lastWaitLog.get('error')).to.eq(err)
             expect(err.message).to.include('`cy.wait()` timed out waiting `100ms` for the 1st request to the route: `getBar`. No request ever occurred.')
 
             done()
@@ -1025,9 +1013,8 @@ describe('src/cy/commands/waiting', () => {
           cy.visit('/fixtures/empty.html')
 
           cy
-          .server()
-          .route(/foo/, {}).as('getFoo')
-          .route(/bar/, {}).as('getBar')
+          .intercept(/foo/, {}).as('getFoo')
+          .intercept(/bar/, {}).as('getBar')
           .window().then((win) => {
             xhrGet(win, '/foo')
 
@@ -1050,8 +1037,7 @@ describe('src/cy/commands/waiting', () => {
 
         it('is a parent command', () => {
           cy
-          .server()
-          .route(/foo/, {}).as('getFoo')
+          .intercept(/foo/, {}).as('getFoo')
           .window().then((win) => {
             xhrGet(win, '/foo')
 
@@ -1066,9 +1052,8 @@ describe('src/cy/commands/waiting', () => {
 
         it('passes as array of referencesAlias', () => {
           cy
-          .server()
-          .route(/foo/, {}).as('getFoo')
-          .route(/bar/, {}).as('getBar')
+          .intercept(/foo/, {}).as('getFoo')
+          .intercept(/bar/, {}).as('getBar')
           .window().then((win) => {
             xhrGet(win, '/foo')
             xhrGet(win, '/bar')
@@ -1077,9 +1062,9 @@ describe('src/cy/commands/waiting', () => {
             return null
           })
           .wait(['@getFoo', '@getBar', '@getFoo']).then(function (xhrs) {
-            const { lastLog } = this
+            const { lastWaitLog } = this
 
-            expect(lastLog.get('referencesAlias')).to.deep.eq([
+            expect(lastWaitLog.get('referencesAlias')).to.deep.eq([
               {
                 name: 'getFoo',
                 cardinal: 1,
@@ -1101,15 +1086,14 @@ describe('src/cy/commands/waiting', () => {
 
         it('#consoleProps waiting on 1 alias', () => {
           cy
-          .server()
-          .route(/foo/, {}).as('getFoo')
+          .intercept(/foo/, {}).as('getFoo')
           .window().then((win) => {
             xhrGet(win, '/foo')
 
             return null
           })
           .wait('@getFoo').then(function (xhr) {
-            expect(this.lastLog.invoke('consoleProps')).to.deep.eq({
+            expect(this.lastWaitLog.invoke('consoleProps')).to.deep.eq({
               Command: 'wait',
               'Waited For': 'getFoo',
               Yielded: xhr,
@@ -1119,9 +1103,8 @@ describe('src/cy/commands/waiting', () => {
 
         it('#consoleProps waiting on multiple aliases', () => {
           cy
-          .server()
-          .route(/foo/, {}).as('getFoo')
-          .route(/bar/, {}).as('getBar')
+          .intercept(/foo/, {}).as('getFoo')
+          .intercept(/bar/, {}).as('getBar')
           .window().then((win) => {
             xhrGet(win, '/foo')
             xhrGet(win, '/bar')
@@ -1129,7 +1112,7 @@ describe('src/cy/commands/waiting', () => {
             return null
           })
           .wait(['@getFoo', '@getBar']).then(function (xhrs) {
-            expect(this.lastLog.invoke('consoleProps')).to.deep.eq({
+            expect(this.lastWaitLog.invoke('consoleProps')).to.deep.eq({
               Command: 'wait',
               'Waited For': 'getFoo, getBar',
               Yielded: [xhrs[0], xhrs[1]], // explicitly create the array here
@@ -1153,8 +1136,7 @@ describe('src/cy/commands/waiting', () => {
           })
 
           cy
-          .server()
-          .route('GET', '*', {}).as('fetch')
+          .intercept('GET', '*', {}).as('fetch')
           .wait('@fetch')
         })
 
@@ -1166,23 +1148,26 @@ describe('src/cy/commands/waiting', () => {
           })
 
           cy
-          .server()
-          .route('GET', '*', {}).as('fetch')
+          .intercept('GET', '*', {}).as('fetch')
           .wait('@fetch', { requestTimeout: 199 })
         })
 
         it('sets default responseTimeout', {
           responseTimeout: 299,
         }, function (done) {
-          cy.on('command:retry', () => {
-            expect(this.lastLog.get('timeout')).to.eq(299)
+          cy.on('command:retry', (command) => {
+            if (command.type === 'response') {
+              expect(this.lastWaitLog.get('timeout')).to.eq(299)
 
-            done()
+              done()
+            }
           })
 
           cy
-          .server({ delay: 100 })
-          .route('GET', '*', {}).as('fetch')
+          .intercept('GET', '*', {
+            body: {},
+            delay: 100,
+          }).as('fetch')
           .window().then((win) => {
             xhrGet(win, '/foo')
 
@@ -1192,15 +1177,19 @@ describe('src/cy/commands/waiting', () => {
         })
 
         it('sets custom responseTimeout', function (done) {
-          cy.on('command:retry', () => {
-            expect(this.lastLog.get('timeout')).to.eq(299)
+          cy.on('command:retry', (command) => {
+            if (command.type === 'response') {
+              expect(this.lastWaitLog.get('timeout')).to.eq(299)
 
-            done()
+              done()
+            }
           })
 
           cy
-          .server({ delay: 100 })
-          .route('GET', '*', {}).as('fetch')
+          .intercept('GET', '*', {
+            body: {},
+            delay: 100,
+          }).as('fetch')
           .window().then((win) => {
             xhrGet(win, '/foo')
 
@@ -1213,11 +1202,11 @@ describe('src/cy/commands/waiting', () => {
           let log
           let retryCount = 0
 
-          cy.on('command:retry', () => {
-            log = log || this.lastLog
+          cy.on('command:retry', (command) => {
+            log = log || this.lastWaitLog
             retryCount++
             if (retryCount === 1) {
-              expect(log.get('timeout')).to.eq(100)
+              expect(log.get('timeout')).to.eq(1000)
 
               // trigger request to move onto response timeout verification
               const win = cy.state('window')
@@ -1225,17 +1214,19 @@ describe('src/cy/commands/waiting', () => {
               xhrGet(win, '/foo')
             }
 
-            if (retryCount === 2) {
-              expect(log.get('timeout')).to.eq(299)
+            if (command.type === 'response') {
+              expect(log.get('timeout')).to.eq(1001)
 
               done()
             }
           })
 
           cy
-          .server({ delay: 100 })
-          .route('GET', '*', {}).as('fetch')
-          .wait('@fetch', { requestTimeout: 100, responseTimeout: 299 })
+          .intercept('GET', '*', {
+            body: {},
+            delay: 100,
+          }).as('fetch')
+          .wait('@fetch', { requestTimeout: 1000, responseTimeout: 1001 })
         })
       })
     })