+++ scripts/after-pack-hook.js
@@ -6,7 +6,8 @@ const os = require('os')
 const path = require('path')
 const { setupV8Snapshots } = require('@tooling/v8-snapshot')
 const { flipFuses, FuseVersion, FuseV1Options } = require('@electron/fuses')
-const { cleanup } = require('./binary/binary-cleanup')
+const { buildEntryPointAndCleanup } = require('./binary/binary-cleanup')
+const { getIntegrityCheckSource, getBinaryEntryPointSource } = require('./binary/binary-sources')
 
 module.exports = async function (params) {
   console.log('****************************')
@@ -55,6 +56,8 @@ module.exports = async function (params) {
   }
 
   if (!['1', 'true'].includes(process.env.DISABLE_SNAPSHOT_REQUIRE)) {
+    await fs.writeFile(path.join(outputFolder, 'index.js'), getBinaryEntryPointSource())
+
     await flipFuses(
       exePathPerPlatform[os.platform()],
       {
@@ -63,7 +66,11 @@ module.exports = async function (params) {
       },
     )
 
-    await setupV8Snapshots(params.appOutDir)
-    await cleanup(outputFolder)
+    // Build out the entry point and clean up prior to setting up v8 snapshots so that the state of the binary is correct
+    await buildEntryPointAndCleanup(outputFolder)
+    await setupV8Snapshots({
+      cypressAppPath: params.appOutDir,
+      integrityCheckSource: getIntegrityCheckSource(outputFolder),
+    })
   }
 }